refactor: split file manager action routes
